+++ classes/output/courseformat/content/sectionselector.php
@@ -64,7 +64,7 @@ class sectionselector extends \core_courseformat\output\local\content\sectionsel
         $numsections = $format->get_last_section_number();
         while ($section <= $numsections) {
             $thissection = $modinfo->get_section_info($section);
-            $url = course_get_url($course, $section, ['navigation' => true]);
+            $url = $format->get_view_url($section, ['navigation' => false]);
             if ($thissection->uservisible && $url && $section != $data->currentsection) {
                 $sectionmenu[$url->out(false)] = get_section_name($course, $section);
             }
@@ -76,6 +76,7 @@ class sectionselector extends \core_courseformat\output\local\content\sectionsel
         $select->formid = 'sectionmenu';
 
         $data->selector = $output->render($select);
+
         return $data;
     }
 }